+++ src/app/src/screens/Container/ScreenHeader.tsx
@@ -1,8 +1,9 @@
-import { IconNames } from "@blueprintjs/icons";
+import { IconName, IconNames } from "@blueprintjs/icons";
 
 // project
 import { Container } from "../../Types";
 import { AppScreenHeader } from "../../components/AppScreenHeader";
+import { pathTo } from "../../Navigator";
 
 import { ActionsMenu } from "./ActionsMenu";
 
@@ -11,13 +12,21 @@ import { ActionsMenu } from "./ActionsMenu";
 interface ScreenHeaderProps {
   container: Container;
   currentScreen: string;
+  listRoutePath?: string;
+  listRouteIcon?: IconName;
 }
 
-export const ScreenHeader: React.FC<ScreenHeaderProps> = ({ container, currentScreen }) => {
+export const ScreenHeader: React.FC<ScreenHeaderProps> = ({ container, currentScreen, listRoutePath, listRouteIcon }) => {
+  let currentListRoutePath = listRoutePath;
+  if (container && !currentListRoutePath) {
+    currentListRoutePath = pathTo("/screens/containers");
+  }
   return (
     <AppScreenHeader
       withBack
       withoutSearch
+      listRoutePath={currentListRoutePath}
+      listRouteIcon={listRouteIcon || IconNames.GRID_VIEW}
       titleIcon={IconNames.BOX}
       titleText={container.Name || container.Id || ""}
       rightContent={<ActionsMenu container={container} expand isActive={(input) => input === currentScreen} />}